requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
langgraph>=0.0.20
openai>=1.3.7
pinecone>=2.2.4
chromadb>=0.4.18
pandas>=2.0.0
numpy>=1.24.0
pdfplumber>=0.10.0  # Python 3.13 compatible PDF processing
pytesseract>=0.3.10
pillow>=10.0.0

pydantic>=2.5.0
typing-extensions>=4.8.0
requests>=2.31.0
aiofiles>=23.2.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2 